+++ src/ai/backend/client/cli/run.py
@@ -1035,26 +1035,26 @@ def start_template(
 
 
 @main.command(aliases=['rm', 'kill'])
-@click.argument('name', metavar='SESSID', nargs=-1)
+@click.argument('session_names', metavar='SESSID', nargs=-1)
 @click.option('-o', '--owner', '--owner-access-key', metavar='ACCESS_KEY',
               help='Specify the owner of the target session explicitly.')
 @click.option('-s', '--stats', is_flag=True,
               help='Show resource usage statistics after termination')
-def terminate(name, owner, stats):
+def terminate(session_names, owner, stats):
     '''
     Terminate the given session.
 
-    SESSID: session ID or its alias given when creating the session.
+    SESSID: session ID given/generated when creating the session.
     '''
-    if len(name) == 0:
+    if len(session_names) == 0:
         print_warn('Specify at least one session ID. Check usage with "-h" option.')
         sys.exit(1)
     print_wait('Terminating the session(s)...')
     with Session() as session:
         has_failure = False
-        for sess in name:
+        for session_name in session_names:
             try:
-                compute_session = session.ComputeSession(sess, owner)
+                compute_session = session.ComputeSession(session_name, owner)
                 ret = compute_session.destroy()
             except BackendAPIError as e:
                 print_error(e)
@@ -1066,16 +1066,52 @@ def terminate(name, owner, stats):
             except Exception as e:
                 print_error(e)
                 has_failure = True
-            if has_failure:
-                sys.exit(1)
         else:
-            print_done('Done.')
+            if not has_failure:
+                print_done('Done.')
             if stats:
                 stats = ret.get('stats', None) if ret else None
                 if stats:
                     print(_format_stats(stats))
                 else:
                     print('Statistics is not available.')
+        if has_failure:
+            sys.exit(1)
+
+
+@main.command()
+@click.argument('session_names', metavar='SESSID', nargs=-1)
+def restart(session_names):
+    '''
+    Restart the given session.
+
+    SESSID: session ID given/generated when creating the session.
+    '''
+    if len(session_names) == 0:
+        print_warn('Specify at least one session ID. Check usage with "-h" option.')
+        sys.exit(1)
+    print_wait('Restarting the session(s)...')
+    with Session() as session:
+        has_failure = False
+        for session_name in session_names:
+            try:
+                compute_session = session.ComputeSession(session_name)
+                compute_session.restart()
+            except BackendAPIError as e:
+                print_error(e)
+                if e.status == 404:
+                    print_info(
+                        'If you are an admin, use "-o" / "--owner" option '
+                        'to terminate other user\'s session.')
+                has_failure = True
+            except Exception as e:
+                print_error(e)
+                has_failure = True
+        else:
+            if not has_failure:
+                print_done('Done.')
+        if has_failure:
+            sys.exit(1)
 
 
 @main.command()